switched time measure from int to float

+++ homcc/monitor/summary.py
@@ -28,22 +28,22 @@ class FileStats:
     """summarized statistics of files"""
 
     filepath: str = field(hash=True)
-    creation_time: int = field(hash=True)
+    creation_time: float = field(hash=True)
 
-    preprocessing_start: Optional[int] = None
-    preprocessing_stop: Optional[int] = None
-    compilation_start: Optional[int] = None
-    compilation_stop: Optional[int] = None
+    preprocessing_start: Optional[float] = None
+    preprocessing_stop: Optional[float] = None
+    compilation_start: Optional[float] = None
+    compilation_stop: Optional[float] = None
 
     def get_compilation_time(self) -> Optional[int]:
         if self.compilation_start is None or self.compilation_stop is None:
             return None
-        return self.compilation_stop - self.compilation_start
+        return int(self.compilation_stop - self.compilation_start)
 
     def get_preprocessing_time(self) -> Optional[int]:
         if self.preprocessing_start is None or self.preprocessing_stop is None:
             return None
-        return self.preprocessing_stop - self.preprocessing_start
+        return int(self.preprocessing_stop - self.preprocessing_start)
 
 
 class SummaryStats:
@@ -57,7 +57,7 @@ class SummaryStats:
         self.host_stats = {}
         self.file_stats = {}
 
-    def register_compilation(self, filename: str, hostname: str, timestamp: int):
+    def register_compilation(self, filename: str, hostname: str, timestamp: float):
         # if new host, add to dict and default its stats
         if hostname not in self.host_stats:
             self.host_stats[hostname] = HostStats(hostname)
@@ -66,10 +66,10 @@ class SummaryStats:
         # track current file stats
         self.file_stats[filename] = FileStats(filename, timestamp)
 
-    def preprocessing_start(self, filename: str, timestamp: int):
+    def preprocessing_start(self, filename: str, timestamp: float):
         self.file_stats[filename].preprocessing_start = timestamp
 
-    def preprocessing_stop(self, filename: str, timestamp: int):
+    def preprocessing_stop(self, filename: str, timestamp: float):
         file_stat = self.file_stats[filename]
         if file_stat.preprocessing_start is None:
             file_stat.preprocessing_start = timestamp
@@ -78,10 +78,10 @@ class SummaryStats:
 
     # for now this would always be preprocessing stop time but this will change in the future once we know when the
     # "Send" phase is over
-    def compilation_start(self, filename: str, timestamp: int):
+    def compilation_start(self, filename: str, timestamp: float):
         self.file_stats[filename].compilation_start = timestamp
 
-    def compilation_stop(self, filename: str, timestamp: int):
+    def compilation_stop(self, filename: str, timestamp: float):
         file_stat = self.file_stats[filename]
         if file_stat.compilation_start is None:
             file_stat.compilation_start = timestamp
@@ -91,7 +91,7 @@ class SummaryStats:
             logger.info("Compilation start timestamp was invalid, assuming zero duration")
         file_stat.compilation_stop = timestamp
 
-    def deregister_compilation(self, filename: str, hostname: str, timestamp: int):
+    def deregister_compilation(self, filename: str, hostname: str, timestamp: float):
         if filename not in self.file_stats:
             logger.info(
                 "Tried to deregister compilation with filename '%s' which was not registered before. Ignoring it.",
